cleanup

+++ main/models.py
@@ -76,7 +76,9 @@ class Phrase(models.Model):
         # help_text="""If you leave this blank, we'll assume it's the current "working on" language.""",
         help_text=""""working on" language if blank""",
     )
-    cosine_similarity = models.FloatField(null=True, blank=False)
+    cosine_similarity = models.DecimalField(
+        null=True, blank=False, max_digits=5, decimal_places=4
+    )
 
     objects = PhraseManager()
